LPS-57513 Move out cluster related classes
