+++ modules/portal-cache/portal-cache-ehcache-cluster/src/com/liferay/portal/cache/ehcache/cluster/internal/distribution/EhcachePortalCacheReplicatorAdapter.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cache.ehcache.internal.distribution;
+package com.liferay.portal.cache.ehcache.cluster.internal.distribution;
 
 import com.liferay.portal.cache.ehcache.event.EhcachePortalCacheListenerAdapter;
 import com.liferay.portal.kernel.cache.PortalCacheReplicator;
